% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoNetworksSimulator.R
\name{simulateModel}
\alias{simulateModel}
\title{Simulate the current trophic model}
\usage{
simulateModel(x)
}
\arguments{
\item{x}{anything}
}
\value{
echos the value of x
}
\description{
Simulate the current trophic model
}

